haskellPackages.base-compat{,batteries}: 0.12.2 -> 0.12.3
Building base-compat-batteries-0.12.2 against OneTuple == 0.4.* as we do
for GHC < 9.0 could lead to confusing results for users of
Data.Tuple.Compat. Thankfully upstream has provided a solution they are
satisfied with and released it as 0.12.3 to which we preemptively
upgrade (but it should enter Stackage LTS 21 soon enough)

See https://github.com/haskell-compat/base-compat/pull/92.

This reverts commit 54ebdad42dd5861538de31319206fcef78ec9df5.